+++ src/Migration/Reporter/MigrationReporterInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\WysiwygModule\Migration\Reporter;
+
+use OxidEsales\WysiwygModule\Migration\DTO\MigrationReportInterface;
+use OxidEsales\WysiwygModule\Migration\DTO\MigrationSummaryInterface;
+
+interface MigrationReporterInterface
+{
+    public function report(MigrationReportInterface $report): MigrationSummaryInterface;
+}